MainForm.BtnJoin.Helper.csを追加 ValidateInputメソッドを新規作成
