+++ src/MJS_fileJoin_src/MJS_fileJoin/MJS_fileJoin.csproj
@@ -52,6 +52,9 @@
     <Compile Include="MainForm.BtnJoin.cs">
       <SubType>Form</SubType>
     </Compile>
+    <Compile Include="MainForm.BtnJoin.Helper.cs">
+      <SubType>Form</SubType>
+    </Compile>
     <Compile Include="MainForm.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -84,7 +87,8 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RemoveSections.cs" />
+    <Compile Include="Word.Merge.RemoveSections.cs" />
+    <Compile Include="WordMerge.CheckIndexSection.cs" />
     <Compile Include="WordMerge.CreateWordDocument.cs" />
     <Compile Include="WordMerge.cs" />
     <Compile Include="WordMerge.UpdateChapterFrontNumbers.cs" />
