+++ src/MJS_fileJoin_src/MJS_fileJoin/MainForm.Utils.cs
@@ -79,114 +79,7 @@ namespace MJS_fileJoin
                 CopyDirectory(copyFrom, copyTo, overwrite);
             }
         }
-        //public static void CopyDirectory(string stSourcePath, string stDestPath, bool bOverwrite)
-        //{
-        //    // コピー先のディレクトリがなければ作成する
-        //    if (!Directory.Exists(stDestPath))
-        //    {
-        //        Directory.CreateDirectory(stDestPath);
-        //        File.SetAttributes(stDestPath, File.GetAttributes(stSourcePath));
-        //        bOverwrite = true;
-        //    }
-
-        //    // コピー元のディレクトリにあるすべてのファイルをコピーする
-        //    if (bOverwrite)
-        //    {
-        //        foreach (string stCopyFrom in Directory.GetFiles(stSourcePath))
-        //        {
-        //            string stCopyTo = Path.Combine(stDestPath, Path.GetFileName(stCopyFrom));
-        //            File.Copy(stCopyFrom, stCopyTo, true);
-        //        }
-
-        //        // 上書き不可能な場合は存在しない時のみコピーする
-        //    }
-        //    else
-        //    {
-        //        foreach (string stCopyFrom in Directory.GetFiles(stSourcePath))
-        //        {
-        //            string stCopyTo = Path.Combine(stDestPath, Path.GetFileName(stCopyFrom));
-
-        //            if (!File.Exists(stCopyTo))
-        //            {
-        //                File.Copy(stCopyFrom, stCopyTo, false);
-        //            }
-        //        }
-        //    }
-
-        //    // コピー元のディレクトリをすべてコピーする (再帰)
-        //    foreach (string stCopyFrom in Directory.GetDirectories(stSourcePath))
-        //    {
-        //        string stCopyTo = Path.Combine(stDestPath, Path.GetFileName(stCopyFrom));
-        //        CopyDirectory(stCopyFrom, stCopyTo, bOverwrite);
-        //    }
-        //}
-
-        //private bool addHtmlDir(string dirPath)
-        //{
-        //    if (Directory.Exists(Path.Combine(Path.GetDirectoryName(dirPath), "headerFile")))
-        //    {
-        //        string[] listFile = Directory.GetFiles(Path.Combine(Path.GetDirectoryName(dirPath), "headerFile"), "???.txt");
-        //        if (listFile.Length == 0)
-        //        {
-        //            MessageBox.Show("「" + Path.Combine(Path.GetDirectoryName(dirPath), "headerFile") + "」に書誌情報ファイルが存在しません。");
-        //        }
-        //        else
-        //        {
-        //            bookInfo[dirPath] = new System.Data.DataTable();
-        //            bookInfo[dirPath].Columns.Add("Column1", typeof(bool));
-        //            bookInfo[dirPath].Columns.Add("Column2", typeof(string));
-        //            bookInfo[dirPath].Columns.Add("Column3", typeof(string));
-        //            bookInfo[dirPath].Columns.Add("Column4", typeof(string));
-        //            bookInfo[dirPath].Columns.Add("Column5", typeof(string));
-
-        //            bool isHtmlDir = false;
-        //            using (StreamReader sr = new StreamReader(listFile[0]))
-        //            {
-        //                while (!sr.EndOfStream)
-        //                {
-        //                    string[] lineStr = (sr.ReadLine()).Split('\t');
-        //                    string htmlName = "";
-        //                    if (lineStr[2].Contains("#"))
-        //                        continue;
-        //                    else
-        //                        htmlName = Path.Combine(dirPath, lineStr[2] + ".html");
-        //                    if (lineStr.Length > 3 && !String.IsNullOrEmpty(lineStr[3]))
-        //                    {
-        //                        bookInfo[dirPath].Rows.Add(true, lineStr[0], lineStr[1], lineStr[2], lineStr[3]);
-        //                    }
-        //                    else
-        //                    {
-        //                        bookInfo[dirPath].Rows.Add(true, lineStr[0], lineStr[1], lineStr[2]);
-
-        //                    }
-
-        //                    if (!isHtmlDir &&
-        //                        File.Exists(htmlName))
-        //                    {
-        //                        isHtmlDir = true;
-        //                    }
-        //                    if (!File.Exists(htmlName)) MessageBox.Show(htmlName + "が存在しません。\r\nwebHelpフォルダ内にHTMLを配置後、結合を実行してください。");
-        //                }
-        //            }
-        //            if (!isHtmlDir)
-        //            {
-        //                MessageBox.Show("書誌情報ファイルに紐づくHTMLファイルが見つかりません。\nフォルダをご確認ください。");
-        //                bookInfo.Remove(dirPath);
-        //            }
-        //            else
-        //            {
-        //                lbHtmlList.Items.Add(dirPath);
-        //                lbHtmlList.SelectedIndex = lbHtmlList.Items.Count - 1;
-        //            }
-        //        }
-        //    }
-        //    else
-        //    {
-        //        MessageBox.Show("「" + Path.Combine(Path.GetDirectoryName(dirPath), "headerFile") + "」フォルダが存在しません。");
-        //    }
-        //    return false;
-        //}
-
+        
         private bool addHtmlDir(string dirPath)
         {
             string headerDir = Path.Combine(Path.GetDirectoryName(dirPath), "headerFile");
